# swagger.yaml
...
openapi: "3.0.0"
info:
  title: "EcoGarden"
  version: "v1"
  description: "This is the API for the EcoGarden project from the Openclassrooms PHP/Symfony course."
paths:
...
  /api/meteo/{city}:
    get:
      summary: "Get weather information for a city"
      security:
        -
          bearerAuth:
      description: "Use a city name or a French postal code"
      parameters:
        - name: "city"
          in: "path"
          required: "true"
          schema:
            type: "string"
            description: "Can be a city name or a French postal code"
      responses:
        200:
          description: "Weather information"
          content:
            application/json:
              schema:
                type: "object"
                description: "Refer to [OpenWeatherMap API documentation](https://openweathermap.org/api)"
        404:
          description: "City not found"
  /api/meteo:
    get:
      summary: "Get weather information for the default city"
      security:
        -
          bearerAuth:
      description: "Uses the city or, if not provided, the zip_code of the authenticated user"
      responses:
        200:
          description: "Weather information"
          content:
            application/json:
              schema:
                type: "object"
                description: "Refer to [OpenWeatherMap API documentation](https://openweathermap.org/api)"
        404:
          description: "City not found"
...
components:
  securitySchemes:
    bearerAuth:
      type: "http"
      scheme: "bearer"
      bearerFormat: "JWT"
...